Making Run_protocol test (#181)
