+++ src/ecdsa/ot_based_ecdsa/triples/batch_random_ot.rs
@@ -371,10 +371,8 @@ mod test {
     use super::*;
     use crate::ecdsa::ot_based_ecdsa::triples::test::run_batch_random_ot;
     use crate::participants::Participant;
-    use crate::protocol::{
-        internal::{make_protocol, Comms},
-        test::run_two_party_protocol,
-    };
+    use crate::protocol::internal::{make_protocol, Comms};
+    use crate::test::run_two_party_protocol;
     use rand_core::OsRng;
 
     #[test]
