+++ src/ecdsa/ot_based_ecdsa/triples/test.rs
@@ -9,10 +9,8 @@ use crate::ecdsa::{Field, Polynomial, ProjectivePoint, Secp256K1ScalarField};
 
 use crate::errors::ProtocolError;
 use crate::participants::Participant;
-use crate::protocol::{
-    internal::{make_protocol, Comms},
-    test::run_two_party_protocol,
-};
+use crate::protocol::internal::{make_protocol, Comms};
+use crate::test::run_two_party_protocol;
 
 /// Create a new triple from scratch.
 ///
