+++ src/ecdsa/robust_ecdsa/test.rs
@@ -9,10 +9,11 @@ use crate::ecdsa::{
     SignatureOption, Tweak,
 };
 use crate::participants::Participant;
-use crate::protocol::{run_protocol, Protocol};
+use crate::protocol::Protocol;
 use crate::test::{
     assert_public_key_invariant, generate_participants, generate_participants_with_random_ids,
-    one_coordinator_output, run_keygen, run_refresh, run_reshare, GenOutput, GenProtocol,
+    one_coordinator_output, run_keygen, run_protocol, run_refresh, run_reshare, GenOutput,
+    GenProtocol,
 };
 
 use rand::rngs::OsRng;
