+++ src/protocol/echo_broadcast.rs
@@ -328,8 +328,8 @@ where
 mod test {
     use super::*;
     use crate::protocol::internal::{make_protocol, Comms};
-    use crate::protocol::{run_protocol, Protocol, ProtocolError};
-    use crate::test::generate_participants;
+    use crate::protocol::{Protocol, ProtocolError};
+    use crate::test::{generate_participants, run_protocol};
 
     /// This function is similar to `do_broadcast` except it is tailored to
     /// consume the inputs instead of borrowing and become suitable for `make_protocol`
