+++ tests/common.rs
@@ -5,9 +5,11 @@ use rand::Rng;
 use rand_core::OsRng;
 
 use threshold_signatures::{
-    self, keygen,
+    self,
+    errors::ProtocolError,
+    keygen,
     participants::Participant,
-    protocol::{run_protocol, Protocol},
+    protocol::{Action, Protocol},
     Ciphersuite, Element, KeygenOutput, Scalar,
 };
 
@@ -20,6 +22,54 @@ pub fn generate_participants(number: u32) -> Vec<Participant> {
 pub fn choose_coordinator_at_random(participants: &[Participant]) -> Participant {
     let index = rand::rngs::OsRng.gen_range(0..participants.len());
     participants[index]
+}
+
+/// Run a protocol to completion, synchronously.
+///
+/// This works by executing each participant in order.
+///
+/// The reason this function exists is as a convenient testing utility.
+/// In practice each protocol participant is likely running on a different machine,
+/// and so orchestrating the protocol would happen differently.
+pub fn run_protocol<T>(
+    mut ps: Vec<(Participant, Box<dyn Protocol<Output = T>>)>,
+) -> Result<Vec<(Participant, T)>, ProtocolError> {
+    let indices: HashMap<Participant, usize> =
+        ps.iter().enumerate().map(|(i, (p, _))| (*p, i)).collect();
+
+    let size = ps.len();
+    let mut out = Vec::with_capacity(size);
+    while out.len() < size {
+        for i in 0..size {
+            while {
+                let action = ps[i].1.poke()?;
+                match action {
+                    Action::Wait => false,
+                    Action::SendMany(m) => {
+                        for j in 0..size {
+                            if i == j {
+                                continue;
+                            }
+                            let from = ps[i].0;
+                            ps[j].1.message(from, m.clone());
+                        }
+                        true
+                    }
+                    Action::SendPrivate(to, m) => {
+                        let from = ps[i].0;
+                        ps[indices[&to]].1.message(from, m);
+                        true
+                    }
+                    Action::Return(r) => {
+                        out.push((ps[i].0, r));
+                        false
+                    }
+                }
+            } {}
+        }
+    }
+
+    Ok(out)
 }
 
 #[allow(clippy::missing_panics_doc)]
