+++ tests/eddsa.rs
@@ -9,10 +9,9 @@ use threshold_signatures::{
     self,
     eddsa::{sign::sign, Ed25519Sha512, SignatureOption},
     participants::Participant,
-    protocol::run_protocol,
 };
 
-use crate::common::GenProtocol;
+use crate::common::{run_protocol, GenProtocol};
 
 type C = Ed25519Sha512;
 type KeygenOutput = threshold_signatures::KeygenOutput<C>;
